Add no_std support
