+++ audio-core/src/lib.rs
@@ -6,6 +6,7 @@
 //!
 //! [audio]: https://docs.rs/audio
 
+#![cfg_attr(not(feature = "std"), no_std)]
 #![deny(missing_docs, rustdoc::broken_intra_doc_links)]
 #![allow(clippy::should_implement_trait)]
