+++ audio-device/examples/wasapi.rs
@@ -4,7 +4,7 @@ use audio_generator::{self as gen, Generator as _};
 
 fn run_output<T>(client: wasapi::Client, mut config: wasapi::ClientConfig) -> Result<()>
 where
-    T: Copy + wasapi::Sample + core::Translate<f32>,
+    T: Copy + wasapi::Sample + audio_core::Translate<f32>,
     [T]: rand::Fill,
 {
     config.sample_rate = 120000;
