+++ audio-device/src/alsa/writer.rs
@@ -1,7 +1,7 @@
+use core::marker;
+
 use crate::alsa::{Error, Pcm, Result};
 use crate::libc as c;
-use core as core;
-use std::marker;
 
 /// A interleaved type-checked PCM writer.
 ///
@@ -30,7 +30,7 @@ impl<'a, T> Writer<'a, T> {
     /// Write an interleaved buffer.
     pub fn write_interleaved<B>(&mut self, mut buf: B) -> Result<()>
     where
-        B: core::Buf<Sample = T> + core::ReadBuf + core::ExactSizeBuf + core::InterleavedBuf,
+        B: audio_core::Buf<Sample = T> + audio_core::ReadBuf + audio_core::ExactSizeBuf + audio_core::InterleavedBuf,
     {
         if buf.channels() != self.channels {
             return Err(Error::ChannelsMismatch {
